new version with error in add user

+++ index.php
@@ -21,8 +21,8 @@ include "db_conn.php";
 
 <body>
   <nav class="navbar navbar-light justify-content-center fs-3 mb-5" style="background-color: #00ff5573;">
-    PHP Complete CRUD Application
-  </nav>
+  Plateforme-Gestion-Projet-Collaboratif  </nav>
+
 
   <div class="container">
     <?php
@@ -43,30 +43,47 @@ include "db_conn.php";
           <th scope="col">First Name</th>
           <th scope="col">Last Name</th>
           <th scope="col">Email</th>
-          <th scope="col">Gender</th>
+          <th scope="col">Role</th>
           <th scope="col">Action</th>
         </tr>
       </thead>
       <tbody>
-        <?php
-        $sql = "SELECT * FROM `crud`";
-        $result = mysqli_query($conn, $sql);
-        while ($row = mysqli_fetch_assoc($result)) {
-        ?>
-          <tr>
-            <td><?php echo $row["id"] ?></td>
-            <td><?php echo $row["first_name"] ?></td>
-            <td><?php echo $row["last_name"] ?></td>
-            <td><?php echo $row["email"] ?></td>
-            <td><?php echo $row["gender"] ?></td>
-            <td>
-              <a href="edit.php?id=<?php echo $row["id"] ?>" class="link-dark"><i class="fa-solid fa-pen-to-square fs-5 me-3"></i></a>
-              <a href="delete.php?id=<?php echo $row["id"] ?>" class="link-dark"><i class="fa-solid fa-trash fs-5"></i></a>
-            </td>
-          </tr>
-        <?php
+      <?php
+include "db_conn.php"; // Include the file with PDO connection
+
+try {
+    $pdo = new PDO("mysql:host=$servername;dbname=$dbname", $username, $password);
+    // Set the PDO error mode to exception
+    $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+
+    // Prepare SQL statement
+    $stmt = $pdo->prepare("SELECT * FROM `crud`");
+    $stmt->execute();
+
+    // Check if there are rows returned
+    if ($stmt->rowCount() > 0) {
+        while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
+            ?>
+            <tr>
+                <td><?php echo $row["id"] ?></td>
+                <td><?php echo $row["first_name"] ?></td>
+                <td><?php echo $row["last_name"] ?></td>
+                <td><?php echo $row["email"] ?></td>
+                <td><?php echo $row["Role"] ?></td>
+                <td>
+                    <a href="edit.php?id=<?php echo $row["id"] ?>" class="link-dark"><i class="fa-solid fa-pen-to-square fs-5 me-3"></i></a>
+                    <a href="delete.php?id=<?php echo $row["id"] ?>" class="link-dark"><i class="fa-solid fa-trash fs-5"></i></a>
+                </td>
+            </tr>
+            <?php
         }
-        ?>
+    } else {
+        echo "<tr><td colspan='6'>No records found.</td></tr>";
+    }
+} catch (PDOException $e) {
+    echo "Failed: " . $e->getMessage();
+}
+?>
       </tbody>
     </table>
   </div>
@@ -74,6 +91,7 @@ include "db_conn.php";
   <!-- Bootstrap -->
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"></script>
 
+  
 </body>
 
 </html>
